Updated to just use a single header
Header.php now displays the appropriate content depending on whether or not logged in. We can delete the other 2 headers once we are happy with this. Still does not refresh on login but code is kind of simpler.
Also updated bloguser to do an alert if we have registering so you get a pop-up of the error message which is easier for the user to spot.